+++ src/Transformers/RegexFilter.php
@@ -29,7 +29,7 @@ class RegexFilter implements Transformer
     /**
      * A pattern to match email addresses.
      *
-     * @var string
+     * @var literal-string
      */
     public const EMAIL = '/[a-z0-9_\-\+\.]+@[a-z0-9\-]+\.([a-z]{2,4})(?:\.[a-z]{2})?/i';
 
@@ -43,56 +43,63 @@ class RegexFilter implements Transformer
     /**
      * The default URL matching pattern.
      *
-     * @var string
+     * @var literal-string
      */
     public const URL = self::GRUBER_1;
 
     /**
      * The original Gruber URL matching pattern.
      *
-     * @var string
+     * @var literal-string
      */
     public const GRUBER_1 = '%\b(([\w-]+://?|www[.])[^\s()<>]+(?:\([\w\d]+\)|([^[:punct:]\s]|/)))%s';
 
     /**
      * The improved Gruber URL matching pattern.
      *
-     * @var string
+     * @var literal-string
      */
     public const GRUBER_2 = '%(?xi)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:\'".,<>?«»“”‘’]))%s';
 
     /**
      * Matches consecutively repeated non word or number characters such as punctuation and special characters.
      *
-     * @var string
+     * @var literal-string
      */
     public const EXTRA_CHARACTERS = '/([^\w\s])(?=[^\w\s]*\1)/';
 
     /**
      * Matches consecutively repeated words.
      *
-     * @var string
+     * @var literal-string
      */
     public const EXTRA_WORDS = '/\b(\w+)(?=\s+\1+\b)/ui';
 
     /**
      * Matches consecutively repeated whitespace characters.
      *
-     * @var string
+     * @var literal-string
      */
     public const EXTRA_WHITESPACE = '/\s(?=\s+)/';
 
     /**
+     * A pattern to match unicode emojis.
+     *
+     * @var literal-string
+     */
+    public const EMOJIS = '/[\x{1F300}-\x{1F5FF}\x{1F900}-\x{1F9FF}\x{1F600}-\x{1F64F}\x{1F680}-\x{1F6FF}\x{2600}-\x{26FF}\x{2700}-\x{27BF}]/u';
+
+    /**
      * A pattern to match Twitter-style mentions (ex. @RubixML).
      *
-     * @var string
+     * @var literal-string
      */
     public const MENTION = '/(@\w+)/';
 
     /**
      * A pattern to match Twitter-style hashtags (ex. #MachineLearning).
      *
-     * @var string
+     * @var literal-string
      */
     public const HASHTAG = '/(#\w+)/';